backend/requirements.txt
# Deep Learning (Python 3.13 compatible)
torch==2.9.0
torchvision==0.24.0
opencv-python==4.10.0.84
pillow==11.0.0

# Geospatial
earthengine-api==1.2.0
shapely==2.0.6
folium==0.18.0

# Data Processing
numpy==2.2.0
pandas==2.2.3
matplotlib==3.9.2
scikit-learn==1.6.0

# API & Web
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.17
python-dotenv==1.0.1
pydantic==2.10.3

# Utilities
requests==2.32.3
tqdm==4.67.1